+++ 黑马微博2期/Classes/MessageCenter(消息)/Controller/HWMessageCenterViewController.m
@@ -58,7 +58,7 @@
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleSubtitle reuseIdentifier:ID];
     }
     
-    cell.textLabel.text = [NSString stringWithFormat:@"test-message-%d", indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"test-message-%d", (int)indexPath.row];
     
     return cell;
 }
